requirements.txt
libvirt-python==8.0.0